(ui) update so that default list does not sort
