+++ invokeai/frontend/web/src/features/nodes/components/sidePanel/WorkflowListMenu/WorkflowSortControl.tsx
@@ -10,6 +10,7 @@ import {
   PopoverContent,
   PopoverTrigger,
   Tooltip,
+  useDisclosure,
 } from '@invoke-ai/ui-library';
 import { useStore } from '@nanostores/react';
 import { $projectId } from 'app/store/nanostores/projectId';
@@ -91,31 +92,39 @@ export const WorkflowSortControl = () => {
     [direction, DIRECTION_OPTIONS]
   );
 
+  const { isOpen, onToggle, onClose } = useDisclosure();
+
   return (
-    <Popover>
-      <PopoverTrigger>
-        <Tooltip label={`Sorting by ${valueOrderBy?.label} ${valueDirection?.label}`}>
-          <IconButton
-            aria-label="Sort Workflow Librar"
-            icon={direction === 'ASC' ? <PiSortAscendingBold /> : <PiSortDescendingBold />}
-            variant="ghost"
-          />
-        </Tooltip>
-      </PopoverTrigger>
-      <PopoverContent>
-        <PopoverBody>
-          <Flex flexDir="column" gap={4}>
-            <FormControl orientation="horizontal" gap={1}>
-              <FormLabel>{t('common.orderBy')}</FormLabel>
-              <Combobox value={valueOrderBy} options={orderByOptions} onChange={onChangeOrderBy} />
-            </FormControl>
-            <FormControl orientation="horizontal" gap={1}>
-              <FormLabel>{t('common.direction')}</FormLabel>
-              <Combobox value={valueDirection} options={DIRECTION_OPTIONS} onChange={onChangeDirection} />
-            </FormControl>
-          </Flex>
-        </PopoverBody>
-      </PopoverContent>
-    </Popover>
+    <>
+      <Tooltip label={`Sorting by ${valueOrderBy?.label} ${valueDirection?.label}`}>
+        <IconButton
+          aria-label="Sort Workflow Librar"
+          icon={direction === 'ASC' ? <PiSortAscendingBold /> : <PiSortDescendingBold />}
+          variant="ghost"
+          onClick={onToggle}
+        />
+      </Tooltip>
+
+      <Popover placement="bottom" isOpen={isOpen} onClose={onClose} returnFocusOnClose={false} closeOnBlur={false}>
+        <PopoverTrigger>
+          <Flex></Flex>
+        </PopoverTrigger>
+
+        <PopoverContent>
+          <PopoverBody>
+            <Flex flexDir="column" gap={4}>
+              <FormControl orientation="horizontal" gap={1}>
+                <FormLabel>{t('common.orderBy')}</FormLabel>
+                <Combobox value={valueOrderBy} options={orderByOptions} onChange={onChangeOrderBy} />
+              </FormControl>
+              <FormControl orientation="horizontal" gap={1}>
+                <FormLabel>{t('common.direction')}</FormLabel>
+                <Combobox value={valueDirection} options={DIRECTION_OPTIONS} onChange={onChangeDirection} />
+              </FormControl>
+            </Flex>
+          </PopoverBody>
+        </PopoverContent>
+      </Popover>
+    </>
   );
 };
